asticaVoice/python/requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
sounddevice>=0.4.6
